+++ dusty/commands/repos.py
@@ -1,4 +1,5 @@
 import os
+import urlparse
 
 from prettytable import PrettyTable
 
@@ -9,6 +10,7 @@ from ..log import log_to_client
 from .. import constants
 from ..payload import daemon_command
 from ..parallel import parallel_task_queue
+from ..systems.known_hosts import ensure_known_hosts
 
 @daemon_command
 def list_repos():
@@ -41,7 +43,6 @@ def _manage_repo(repo_name):
     else:
         log_to_client('No overriden repos found by name {}'.format(repo_name))
 
-
 @daemon_command
 def manage_repo(repo_name):
     _manage_repo(repo_name)
@@ -60,18 +61,34 @@ def override_repos_from_directory(source_path):
         if os.path.isdir(repo_path):
             override_repo(repo.remote_path, repo_path)
 
+def add_known_hosts_for_repos(repos):
+    hosts = set()
+    for repo in repos:
+        assembled_remote_path = repo.assemble_remote_path()
+        if assembled_remote_path.startswith('ssh://'):
+            parsed = urlparse.urlparse(assembled_remote_path)
+            if not parsed.hostname:
+                raise RuntimeError('Unable to parse hostname from repo {}'.format(repo.remote_path))
+            hosts.add(parsed.hostname)
+    ensure_known_hosts(hosts)
+
+def update_specs_repo_and_known_hosts():
+    specs_repo = get_specs_repo()
+    if not specs_repo.is_overridden:
+        log_to_client('Updating managed copy of specs-repo before loading specs')
+        add_known_hosts_for_repos([specs_repo])
+        specs_repo.update_local_repo()
+    add_known_hosts_for_repos(get_all_repos(active_only=True, include_specs_repo=False))
+
 @daemon_command
 def update_managed_repos(force=False):
     """For any active, managed repos, update the Dusty-managed
     copy to bring it up to date with the latest master."""
     log_to_client('Pulling latest updates for all active managed repos:')
-    overrides = set(get_config_value(constants.CONFIG_REPO_OVERRIDES_KEY))
-    specs_repo = get_specs_repo()
-    if not specs_repo.is_overridden:
-        log_to_client('Updating managed copy of specs-repo before loading specs')
-        specs_repo.update_local_repo()
+    update_specs_repo_and_known_hosts()
+    repos_to_update = get_all_repos(active_only=True, include_specs_repo=False)
     with parallel_task_queue() as queue:
         log_to_client('Updating managed repos')
-        for repo in get_all_repos(active_only=True, include_specs_repo=False):
+        for repo in repos_to_update:
             if not repo.is_overridden:
                 repo.update_local_repo_async(queue, force=force)